Adding 1st SimulationStat
