+++ src/main/java/pl/edu/pwr/bum/simulation/random/events/RandomEventHandler.java
@@ -1,0 +1,4 @@
+package pl.edu.pwr.bum.simulation.random.events;
+
+public class RandomEventHandler {
+}
