+++ src/main/java/pl/edu/pwr/bum/simulation/stats/locations/ActionField.java
@@ -1,6 +1,8 @@
 package pl.edu.pwr.bum.simulation.stats.locations;
 
 
+import pl.edu.pwr.bum.simulation.app.SimulationStats;
+
 public class ActionField implements Location{
     private final SimulationStats.FieldType fieldType = SimulationStats.FieldType.ACTION_FIELD;
     private enum ActionType {}
