--- src/main/java/pl/edu/pwr/bum/simulation/stats/locations/StatsForSingleTurn.java
@@ -1,5 +1,0 @@
-package pl.edu.pwr.bum.simulation.stats.locations;
-
-public class StatsForSingleTurn {
-    private Location location ;
-}
